refactor(gui): extract useWorktreeTaskTitles to its own module
Exporting the hook from the panel component file tripped
react-refresh/only-export-components (CI lint runs --max-warnings 0).
Move the hook and its private helpers into use-worktree-task-titles.ts;
the panel and the unit test import it from there. No behavior change.
