+++ clients/gui-app/src/components/settings/panels/worktrees-settings-panel.tsx
@@ -14,11 +14,7 @@ import {
   type ReactNode,
   type RefCallback,
 } from "react";
-import {
-  useQueryClient,
-  type QueryClient,
-  type UseQueryResult,
-} from "@tanstack/react-query";
+import { useQueryClient, type QueryClient } from "@tanstack/react-query";
 import { useNavigate } from "@tanstack/react-router";
 import { defaultRangeExtractor, useVirtualizer } from "@tanstack/react-virtual";
 import { toast } from "sonner";
@@ -49,11 +45,6 @@ import type {
   WorktreeHostEntry,
   WorktreeHostEntryV14,
 } from "@traycer/protocol/host/index";
-import {
-  GET_TASK_CONTEXTS_MAX_IDS,
-  type GetTaskContextsResponse,
-  type ListTaskLight,
-} from "@traycer/protocol/host/epic/unary-schemas";
 import type {
   WorktreeEntryScripts,
   WorktreePrState,
@@ -71,7 +62,6 @@ import {
 } from "@traycer-clients/shared/worktree/classify-worktree";
 import type { HostClient } from "@traycer-clients/shared/host-client/host-client";
 import type { HostDirectoryEntry } from "@traycer-clients/shared/host-client/host-directory";
-import type { HostRpcError } from "@traycer-clients/shared/host-transport/host-messenger";
 import {
   buildTaskMergeRollups,
   taskMergeRollupEqual,
@@ -119,14 +109,12 @@ import { useReactiveActiveHostId } from "@/hooks/host/use-reactive-active-host-i
 import { useHostDirectoryList } from "@/hooks/host/use-host-directory-list-query";
 import { useHostReachability } from "@/hooks/agent/use-host-reachability";
 import { useHostClientFor } from "@/hooks/host/use-host-client-for";
-import { useHostQueries } from "@/hooks/host/use-host-queries";
 import { useClipboardCopy } from "@/hooks/ui/use-clipboard-copy";
 import { useWorktreeDeleteStreamTransportFactory } from "@/lib/host/use-worktree-delete-stream-transport";
 import type { DurableStreamTransport } from "@/lib/host/durable-stream-transport";
 import { useRefreshSpinner } from "@/hooks/use-refresh-spinner";
 import { useRelativeTimestamp } from "@/lib/relative-time";
-import { useCloudEpicTasksQuery } from "@/hooks/epics/use-cloud-epic-tasks-query";
-import { readEpicTitlesFromCloudTaskCaches } from "@/lib/cloud-epic-tasks-query/cache";
+import { useWorktreeTaskTitles } from "./use-worktree-task-titles";
 import { invalidateWorktreeListingAndBindingCaches } from "@/hooks/worktree/invalidations";
 import {
   backgroundForegroundWorktreeDeleteForHost,
@@ -164,7 +152,6 @@ type WorktreeTierFilterSet = ReadonlySet<WorktreeTier>;
 const EMPTY_TIER_FILTER: WorktreeTierFilterSet = new Set();
 const WORKTREES_REFRESH_TIMEOUT_MS = 10_000;
 const EMPTY_REPO_KEY_SET: ReadonlySet<string> = new Set();
-const EMPTY_TASK_TITLES: ReadonlyMap<string, string> = new Map();
 
 // Virtualization tuning. Row/header heights are estimates only - each rendered
 // item is measured (`virtualizer.measureElement`) so variable-height rows (a
@@ -709,150 +696,6 @@ function WorktreesPartialListingBanner(props: {
         className="h-auto shrink-0 p-0 text-current"
       />
     </div>
-  );
-}
-
-/**
- * Resolves each owner `epicId` on the listing to its Task title in two tiers:
- *
- * 1. **Tier 1 (free):** scan cloud epic-tasks caches the app already maintains
- *    for the signed-in user (`readEpicTitlesFromCloudTaskCaches`). Scope is
- *    `hostId: null` so an epic cached under any host still resolves. We warm
- *    those caches with the shared first-page query History/home use.
- * 2. **Tier 2:** one (or cap-sized) `epic.getTaskContexts` batch for the
- *    still-unresolved ids, keyed by `(hostId, userId, sorted id set)` via the
- *    host-query wrapper. `null` (deleted / not permitted) stays absent → the
- *    chip renderer keeps the muted "Owner unresolved" demotion. Older hosts
- *    that lack the method fail with `E_HOST_UNSUPPORTED` and we fall back to
- *    tier 1 only (today's behavior). Generic/network errors leave rows
- *    unresolved without a new failure surface.
- */
-export function useWorktreeTaskTitles(
-  client: HostClient<HostRpcRegistry> | null,
-  worktrees: readonly WorktreeHostEntryV14[],
-): ReadonlyMap<string, string> {
-  const queryClient = useQueryClient();
-  const epicIds = useMemo(
-    () =>
-      [
-        ...new Set(
-          worktrees.flatMap((entry) =>
-            entry.owners.map((owner) => owner.epicId),
-          ),
-        ),
-      ].sort((left, right) => left.localeCompare(right)),
-    [worktrees],
-  );
-  // Only warm the shared cloud-tasks cache when something actually needs a title.
-  const cloud = useCloudEpicTasksQuery(undefined, {
-    enabled: epicIds.length > 0,
-  });
-  const userId = cloud.currentUserId;
-  const cloudTasks = cloud.tasks;
-
-  const tier1Titles = useMemo(() => {
-    if (userId === null || epicIds.length === 0) return EMPTY_TITLE_RECORD;
-    // `cloudTasks` is a recompute trigger: the read scans the query cache
-    // directly, so we re-derive whenever a fetched page changes it.
-    void cloudTasks;
-    return readEpicTitlesFromCloudTaskCaches(
-      queryClient,
-      { hostId: null, userId },
-      epicIds,
-    );
-  }, [queryClient, userId, epicIds, cloudTasks]);
-
-  const unresolvedIds = useMemo(
-    () => epicIds.filter((id) => !Object.hasOwn(tier1Titles, id)),
-    [epicIds, tier1Titles],
-  );
-
-  const batchRequests = useMemo(
-    () =>
-      chunkTaskIds(unresolvedIds, GET_TASK_CONTEXTS_MAX_IDS).map((taskIds) => ({
-        method: "epic.getTaskContexts" as const,
-        params: { taskIds: [...taskIds] },
-      })),
-    [unresolvedIds],
-  );
-
-  const tier2Titles = useHostQueries<
-    HostRpcRegistry,
-    "epic.getTaskContexts",
-    ReadonlyMap<string, string>
-  >({
-    client,
-    requests: batchRequests,
-    cacheKeyIdentity: userId === null ? undefined : userId,
-    options: {
-      enabled: userId !== null && unresolvedIds.length > 0,
-    },
-    combine: combineTaskContextTitleResults,
-  });
-
-  return useMemo(() => {
-    if (userId === null || epicIds.length === 0) return EMPTY_TASK_TITLES;
-    if (tier2Titles.size === 0) {
-      return new Map(Object.entries(tier1Titles));
-    }
-    const merged = new Map(Object.entries(tier1Titles));
-    for (const [id, title] of tier2Titles) {
-      if (!merged.has(id)) merged.set(id, title);
-    }
-    return merged;
-  }, [userId, epicIds, tier1Titles, tier2Titles]);
-}
-
-const EMPTY_TITLE_RECORD: Record<string, string> = {};
-
-function combineTaskContextTitleResults(
-  results: Array<UseQueryResult<GetTaskContextsResponse, HostRpcError>>,
-): ReadonlyMap<string, string> {
-  // Older host: method unsupported → stay on tier 1 only (no throw).
-  if (
-    results.some((result) => result.error?.code === "E_HOST_UNSUPPORTED")
-  ) {
-    return EMPTY_TASK_TITLES;
-  }
-  const titles = new Map<string, string>();
-  for (const result of results) {
-    if (result.data === undefined) continue;
-    for (const title of titlesFromTaskContextsResponse(result.data)) {
-      if (!titles.has(title.id)) titles.set(title.id, title.title);
-    }
-  }
-  return titles;
-}
-
-function titlesFromTaskContextsResponse(
-  response: GetTaskContextsResponse,
-): ReadonlyArray<{ readonly id: string; readonly title: string }> {
-  return Object.values(response.tasks).flatMap((task) => {
-    const extracted = titleFromListTaskLight(task);
-    return extracted === null ? [] : [extracted];
-  });
-}
-
-function titleFromListTaskLight(
-  task: ListTaskLight | null,
-): { readonly id: string; readonly title: string } | null {
-  if (task === null) return null;
-  const light = task.epic?.light;
-  if (light === null || light === undefined) return null;
-  const title = light.title.trim();
-  if (title.length === 0) return null;
-  return { id: light.id, title };
-}
-
-function chunkTaskIds(
-  ids: readonly string[],
-  maxPerChunk: number,
-): ReadonlyArray<ReadonlyArray<string>> {
-  if (ids.length === 0) return [];
-  return Array.from(
-    { length: Math.ceil(ids.length / maxPerChunk) },
-    (_value, index) =>
-      ids.slice(index * maxPerChunk, (index + 1) * maxPerChunk),
   );
 }
